requirements.txt
django==5.1.1
pydantic==2.9.2
django-ninja==1.3.0
django-ninja-extra==0.21.4
requests==2.32.3
python-dotenv==1.0.1
dj-database-url==2.2.0
psycopg2-binary==2.9.9
